requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
astroid==2.4.0            # via pylint
attrs==19.3.0             # via -r requirements.in, jsonschema, pytest
certifi==2020.4.5.1       # via requests
chardet==3.0.4            # via requests
click==7.1.1              # via flask
environs==7.4.0           # via -r requirements.in
flask==1.1.2              # via -r requirements.in
gunicorn==20.0.4          # via -r requirements.in
idna==2.9                 # via requests, tldextract
isort==4.3.21             # via pylint
itsdangerous==1.1.0       # via flask
jinja2==2.11.2            # via flask
jsonschema==3.2.0         # via -r requirements.in
lazy-object-proxy==1.4.3  # via astroid
markupsafe==1.1.1         # via jinja2
marshmallow==3.5.1        # via environs
mccabe==0.6.1             # via pylint
more-itertools==8.2.0     # via pytest
packaging==20.3           # via pytest
pluggy==0.13.1            # via pytest
prometheus_client==0.7.1  # via -r requirements.in
py==1.8.1                 # via pytest
pylint==2.5.0             # via -r requirements.in
pyparsing==2.4.7          # via packaging
pyrsistent==0.16.0        # via jsonschema
pytest==5.4.1             # via -r requirements.in
python-dotenv==0.13.0     # via environs
redis==3.4.1              # via -r requirements.in
requests-file==1.5.0      # via tldextract
requests==2.23.0          # via requests-file, tldextract
six==1.14.0               # via astroid, jsonschema, packaging, pyrsistent, requests-file
tldextract==2.2.2         # via -r requirements.in
toml==0.10.0              # via pylint
urllib3==1.25.9           # via requests
wcwidth==0.1.9            # via pytest
werkzeug==1.0.1           # via flask
wrapt==1.12.1             # via astroid

# The following packages are considered to be unsafe in a requirements file:
# setuptools
